refactor(tuirealm): move "AttrValue" to its own file
